+++ packages/form/src/components/FieldSet/demos/form-fieldset.tsx
@@ -56,7 +56,7 @@ export default () => {
             },
           },
         ]}
-        transform={(value: any) => ({ startTime: value[0], endTime: value[1] })}
+        transform={(value: any) => ({ list: value, startTime: value[0], endTime: value[1] })}
       >
         <ProFormText width="md" />
         <ProFormText width="md" />
@@ -66,7 +66,7 @@ export default () => {
       <ProFormFieldSet
         name="list"
         label="组件列表"
-        transform={(value: any) => ({ startTime: value[0], endTime: value[1] })}
+        transform={(value: any) => ({ list: value, startTime: value[0], endTime: value[1] })}
       >
         <ProFormText width="md" readonly />
         -