+++ packages/provider/src/locale/ko_KR.tsx
@@ -61,6 +61,11 @@ export default {
       save: '저장',
       cancel: '취소',
       delete: '삭제',
+      add: '데이터 행 추가',
     },
+  },
+  switch: {
+    open: '열',
+    close: '가까 운',
   },
 };