+++ packages/provider/src/locale/ms_MY.tsx
@@ -1,5 +1,13 @@
 export default {
   moneySymbol: 'RM',
+  form: {
+    lightFilter: {
+      more: 'Lebih banyak',
+      clear: 'Jelas',
+      confirm: 'Mengesahkan',
+      itemUnit: 'Item',
+    },
+  },
   tableForm: {
     search: 'Cari',
     reset: 'Menetapkan semula',
@@ -40,7 +48,24 @@ export default {
     densityMiddle: 'Tengah',
     densitySmall: 'Kecil',
   },
+  stepsForm: {
+    next: 'Seterusnya',
+    prev: 'Sebelumnya',
+    submit: 'Selesai',
+  },
   loginForm: {
     submitText: 'Log Masuk',
+  },
+  editableTable: {
+    action: {
+      save: 'Simpan',
+      cancel: 'Membatalkan',
+      delete: 'Menghapuskan',
+      add: 'tambah baris data',
+    },
+  },
+  switch: {
+    open: 'Terbuka',
+    close: 'Tutup',
   },
 };