+++ tests/layout/mobile.test.tsx
@@ -1,5 +1,6 @@
 ﻿import { BasicLayout } from '@ant-design/pro-components';
-import { mount, render } from 'enzyme';
+import { render as reactRender } from '@testing-library/react';
+import { render } from 'enzyme';
 import { act } from 'react-dom/test-utils';
 import { waitForComponentToPaint } from '../util';
 import defaultProps from './defaultProps';
@@ -100,11 +101,11 @@ describe('mobile BasicLayout', () => {
 
   it('📱 layout collapsedButtonRender', async () => {
     const onCollapse = jest.fn();
-    const html = mount(
+    const html = reactRender(
       <BasicLayout
         {...defaultProps}
         onCollapse={onCollapse}
-        collapsed
+        collapsed={false}
         collapsedButtonRender={() => {
           return 'div';
         }}
@@ -115,11 +116,13 @@ describe('mobile BasicLayout', () => {
 
     waitForComponentToPaint(html);
     act(() => {
-      html.find('span.ant-pro-global-header-collapsed-button').simulate('click');
+      html.baseElement
+        ?.querySelector<HTMLSpanElement>('span.ant-pro-global-header-collapsed-button')
+        ?.click();
     });
     waitForComponentToPaint(html);
     act(() => {
-      html.find('div.ant-drawer-mask').simulate('click');
+      html.baseElement?.querySelector<HTMLDivElement>('div.ant-drawer-mask')?.click();
     });
     waitForComponentToPaint(html);
     expect(onCollapse).toHaveBeenCalled();